Removes use of `MetaIntegration`, `HTTPIntegration` and kinesis
`agent` and AWS kinesis hasn't been used for years.

Removes `MetaIntegration`. It was a idea to use a python metaclass to eliminate
the use of a factory function. The down sides of this were poor typing and lack
of clarity.

Removes `HTTPIntegration`. It was a class that used `MetaIntegration`,
configured a HTTP endpoint and had retry. This included confusing assocation of
the service name in config.py and the property `service_name` in the
instantiating classes meta class. We are no longer using `MetaIntegration` and
we are using google's retry code.