Change the base class of EndorsementRequestsAudit, EndorsementAudit to Base.
(and some others for same reason)

This is an incorrect use of model inheritance.

When you inherit a model, you create rows to both tables.

If you add EndorsementAudit, you also add Endorsement record. IOW, you cannot treat these as separate entity. This would be only correct if you tell SQLAlchemy to ues polymorphics and specify the mapper args to all of participating classes. I don't think we want to do this.