Deal with gcld3 problems
Makes it so tests and code work without gcld3 and protobuff-compile.  Makes
gcld3 extra in pyproject.toml. Makes tests that use gcld3 pytest.skip if they
cannot do imports.

At some point gcld3 was added to the requirements but it needs protobuff-compile
which cannot be installed with pip. This caused problems with users of
arxiv-base that didn't need gcld3.

metadata.py needs gcld3.