WIP adds a new jwt.py file which makes a NG style jwt
The intent of this is to have much the JWT making and decoding code in one place
so the shape of the JWT and how it is encoded is in one place. In NG the way a
had the code unnecessarily split over sever files making it difficult understand
and reuse.